requirements.txt
XlsxWriter
python-slugify
nltk==3.*